requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
matplotlib==3.8.2
networkx==3.2.1
Werkzeug==3.0.1
nltk==3.8.1
pandas==2.1.4
Pillow==10.1.0
pytesseract==0.3.10
PyMuPDF==1.23.8